Fixed the InterSenseModule. Now it delivers the correct angles returned by the InertiaCube (before, there was a problem with the _real_ component of the quaternion).
IMPORTANT:
The angles returned by the Intersense sensors will now be different. They have now been fixed, however, if your application does not work anymore is maybe because you found a hack to get around this problem. You should be able to get rid of this hack now. 
I also introduced a flag for the ResetHeading call of intersense, it is at the moment set to true by default for legacy issues. However most of the applications will have to set it to "false". Leave it with its default value only if you know what you're doing.
I also introduced a "HeadTracking" option, where I align the axes of the inertia cube with that of inventor, plus I change from Left to Right handed coordinate system. 
I also introduced a "verbose" flag for a quieter connection to the sensors. Plus, those annoying warnings for none connected trackers were removed.